requirements.txt
Django==5.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.5.*
django-cors-headers==4.6.*
django-filter==24.3
django-cleanup==9.0.0
django-storages==1.14.6
django-redis==6.0.0

gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8

celery==5.6.*
redis[hiredis]==5.2.*

web3==6.0.0
boto3==1.37.16

solana==0.36.6
solders==0.26.0
base58==2.1.1
PyNaCl==1.5.0